graceful degradation for ie background
Strangely, for IE 11, the background image only shows when you either hit the "projects" button or the "to the top" caret.  Hard coding an image tag inline overrides the css method of background for the other browsers.  Setting overflow to auto causes it's own set of problems.  In the end, graceful degradation was decided upon by adding a background color that matches the color of the image.